[dt] Added creator functions for two time steppers
